SortedIntervals

Documentation for SortedIntervals.

SortedIntervals._intervals_are_ordered

SortedIntervals.clip_interval_duration

SortedIntervals.interval_indices

SortedIntervals.interval_intersections

SortedIntervals.intervals_diff

SortedIntervals.join_intervals

SortedIntervals.join_intervals!

SortedIntervals.parse_ranges_str

SortedIntervals.throttle

SortedIntervals._intervals_are_ordered — Method

Checks that each interval is well ordered, the set of intervals is sorted, and the intervals are non-overlapping. Returns a boolean.

SortedIntervals.clip_interval_duration — Method

Clip an interval while trying to maintain its duration

SortedIntervals.interval_indices — Method

interval_indices(
    basis::Union{<:AbstractVector, AbstractRange}, start::Number, stop::Number
) -> i_b, i_e

Find the indices in basis that correspond to the interval specified by start and stop.

SortedIntervals.interval_intersections — Method

Assumes each list is sorted and non-overlapping

SortedIntervals.intervals_diff — Method

intervals_diff(ints_a, ints_b) -> ints_out

Find the 'set diff' of the intervals in ints_a and ints_b. Assumes both inputs are sorted.

SortedIntervals.join_intervals! — Function

join_intervals!(ints::Vector{NTuple{2, <:Number}}, min_gap)

Join a list of sorted intervals, ints, if the gap between successive intervals is less than min_gap. Mutates input in-place

Assumes ints are sorted by their first index.

SortedIntervals.join_intervals — Method

join_intervals(f, ints::Vector{NTuple{2, <:Number}}, min_gap)

Like join_intervals!, but does not mutate input.

SortedIntervals.parse_ranges_str — Method

parse_ranges_str(s::AbstractString)

Parse strings like "3", "1-5", or "1, 2-10, 12-30" into a vector of the integers in specified ranges.

SortedIntervals.throttle — Method

throttle(xs::AbstractVector{T}, min_gap::Number) where T<:Number ->
Vector{NTuple{2, T}}

Join points in x into ranges, if the difference between neighboring elements is less than min_gap. Assumes x is sorted.